Exported drop down data fetching functions to util file to be re-used
