+++ front-end/src/components/EventForm/apiFetches.ts
@@ -1,14 +1,6 @@
 // type imports
 import {
-	GetMeetingTypesResponse,
-	GetRsoResponse,
-	GetStatesResponse,
-	GetUniversitiesResponse
-} from "../../types/apiResponseBodies";
-import {
 	CreateEventRequest,
-	GetRsoRequest,
-	GetUniversitiesRequest,
 	UpdateEventRequest
 } from "../../types/apiRequestBodies";
 
@@ -16,6 +8,12 @@ import {
 import buildpath from "../../Utils/buildpath";
 import { FormState } from "./componentSetup";
 import { Event } from "src/types/eventTypes";
+import {
+	fetchStates,
+	fetchAllRSOs,
+	fetchMeetingTypes,
+	fetchUniversityData,
+} from "../../Utils/apiDropDownData";
 
 export interface DropDownData
 {
@@ -23,124 +21,6 @@ export interface DropDownData
 	RSOs: Map<string, number>,
 	meetingTypes: Map<string, number>,
 	universities: Map<string, number>,
-};
-
-const fetchStates = async (): Promise<Map<string, number>> => {
-
-	let request: Object = {
-		method: "POST",
-		body: JSON.stringify({}),
-		headers: {
-			"Content-Type": "application/json",
-		},
-	};
-
-	let response: GetStatesResponse = (await (await fetch(buildpath("/api/getStates"), request)).json());
-	let mappedStates: Map<string, number> = new Map<string, number>();
-
-	if (!response.success)
-	{
-		console.error(response.error);
-	}
-
-	for (let state of response.states)
-	{
-		mappedStates.set(state.name, state.ID);
-	}
-
-	return mappedStates;
-};
-
-const fetchAllRSOs = async (universityID: number): Promise<Map<string, number>> => {
-	if (typeof universityID === "string") {
-		console.warn("UniversityID is not numeric");
-	}
-
-	let payload: GetRsoRequest = {
-		// TODO: Remove hard coded values when app is functional
-		universityID: (typeof universityID === "string") ? 1 : universityID,
-	};
-
-	let request: Object = {
-		method: "POST",
-		body: JSON.stringify(payload),
-		headers: {
-			"Content-Type": "application/json",
-		},
-	};
-
-	let response: GetRsoResponse = await (await fetch(buildpath("/api/getRso"), request)).json();
-	let mappedRSOs: Map<string, number> = new Map<string, number>();
-
-	if (!response.success)
-	{
-		console.error(response.error);
-	}
-
-	for (let rso of response.RSOs)
-	{
-		mappedRSOs.set(rso.name, rso.ID);
-	}
-
-	return mappedRSOs;
-};
-
-const fetchMeetingTypes = async (): Promise<Map<string, number>> => {
-	let request: Object = {
-		method: "POST",
-		body: JSON.stringify({}),
-		headers: {
-			"Content-Type": "application/json",
-		},
-	};
-
-	let response: GetMeetingTypesResponse = await (await fetch(buildpath("/api/getMeetingTypes"), request)).json();
-	let mappedTypes: Map<string, number> = new Map<string, number>();
-
-	if (!response.success)
-	{
-		console.error(response.error);
-	}
-
-	for (let type of response.meetingTypes)
-	{
-		mappedTypes.set(type.name, type.ID);
-	}
-
-	return mappedTypes;
-};
-
-const fetchUniversityData = async (universityID: number): Promise<Map<string, number>> => {
-	let payload: GetUniversitiesRequest = {};
-
-	if (typeof universityID === "string") {
-		console.warn("Stud User university ID is not specified");
-	} else {
-		payload.schoolID = universityID;
-	}
-
-	let request: Object = {
-		method: "POST",
-		body: JSON.stringify(payload),
-		headers: {
-			"Content-Type": "application/json",
-		},
-	};
-
-	let response: GetUniversitiesResponse = await (await fetch(buildpath("/api/getUniversities"), request)).json();
-	let mappedUniversities: Map<string, number> = new Map<string, number>();
-
-	if (!response.success)
-	{
-		console.error(response.error);
-	}
-
-	for (let university of response.universities)
-	{
-		mappedUniversities.set(university.name, university.ID);
-	}
-
-	return mappedUniversities;
 };
 
 // fetch data from the api
